feat: 수정된 API 동작 테스트(#12)

+++ src/main/java/reviewme/be/feedback/request/PostFeedbackRequest.java
@@ -25,5 +25,5 @@ public class PostFeedbackRequest {
 
     @Schema(description = "이력서 페이지", example = "1")
     @NotNull(message = "이력서의 몇 페이지에 질문을 하는 지는 필수 입력 값입니다.")
-    private Long resumePage;
+    private Integer resumePage;
 }
